services/api/requirements.txt
fastapi[standard]==0.116.1
pydantic-settings==2.10.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.43
alembic==1.16.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.3.0
